+++ tests/testthat/test-get_er_from_jags.R
@@ -9,8 +9,8 @@ test_that("The function produces a ranked matrix", {
                            id_application = "application",
                            id_voter = "voter",
                            grade_variable = "num_grade",
-                           n_chains = 2, n_iter = 1000,
-                           n_burnin = 1000)
+                           n_chains = 4, n_adapt = 10000, n_iter = 10000,
+                           n_burnin = 10000)
   expect_equal(test$rankings %>%
                    nrow(), n_application)
 })